+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "kubera-ui",
   "license": "Apache-2.0",
-  "version": "0.1.2",
+  "version": "0.1.4",
   "private": false,
   "description": "Component Library for Kubera products",
   "author": "MayaData, Inc.",
@@ -51,7 +51,7 @@
   "module": "lib.esm.js",
   "types": "index.d.ts",
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
+    "@material-ui/core": "^4.10.2",
     "@material-ui/icons": "^4.9.1",
     "react": "^16.9.0",
     "react-dom": "^16.9.0"
