--- src/core/Input/InputField/index.ts
@@ -1,1 +1,0 @@
-export * from './InputField';
