--- src/core/Input/base.ts
@@ -1,3 +1,0 @@
-import { InputBaseProps } from '@material-ui/core/InputBase';
-
-export type BaseInputProps = Omit<InputBaseProps, 'variant'>;
