+++ src/core/InputField/base.ts
@@ -1,0 +1,3 @@
+import { TextFieldProps } from '@material-ui/core/TextField';
+
+export type BaseInputProps = Omit<TextFieldProps, 'variant'>;
