+++ src/core/index.ts
@@ -1,2 +1,2 @@
 export * from './Button';
-export * from './Input';
+export * from './InputField';
